Initial Rust based AX5043 driver, examples
This is a first pass at a Rust AX5043 driver. The high level design
intent is a two layer driver, the lower layer closely mirroring the SPI
interface (lib.rs), making it convenient to copy from the datasheet and
talk to the EE types about, and a higher layer (fifo.rs/config.rs) that
tries to hide the complexity and quirks of the chip, making things
easier for RF types to deal with.

The lower layer is fine enough, but to really realize what I wanted I'd
need to write custom derive macros (taking a lot of inspiration from
tock_registers) but I'm not yet very learned in how to write those so
it's currently only stubbed out in the ax5043_derive subcrate.