Add argument handling to runner.sh
The runner.sh now passes program arguments from the cargo invocation to
the program being run instead of silently ignoring them.

This patch also runs the program with RUST_BACKTRACE set and also
sketches out how to make the runner more flexible for different targets,
but it's not a complete solution at the moment, it would ideally depend
on how the target flag is invoked or something.