Split tui widgets into library part
So that they can be reused, for example in the new tuitx example.

Implementing the widget directly on the register struct makes sense in a
Rust sort of way, I didn't do it at first because in other languages
keeping ui and data separate means like different classes or sections.
This is like that, but not.