Register type fixes for FIFODATA, TRKFREQ, RSSIREFERENCE
TRKKFREQ and RSSIREFERENCE both appear to be signed. FIFODATA appears to
only be happy when read in full chunk lengths, it returns ~mystery data~
when trying to read a chunk in two parts. This is inconvenient because
the only way to know how much to read is from FIFOCOUNT now, and we have
to handle reading multiple chunks in one go instead of one at a time.