Refactor Registers, errors
This leverages a macro to change the way register io types work, leading
to increased flexibility especially concerning lifetimes. It also
introduces thiserror and anyhow for ease of error handling, it makes
use of Result much more smooth.

Previously each register in Registers would be a struct that contained
references to all the necessary resources to do IO. This meant that they
all had to be non-mut and that they were repeated hundreds of times.
This especially caused the callback to be pretty restricted by only
being Fn instead of FnMut. The new way is to have the resources be owned
by the Registers struct and be borrowed each time a register needs to be
accessed, constructing the previous struct at access time. This means
that the resources are borrowed only once and can be done mutably if
needed. The downside is that each register in Registers is a method now
instead of a field.

While I was there reading and writing registers were turned into a trait
to reduce duplication of the read and write methods.

The examples were all updated to use the new interface, and a few no
longer needed examples were removed.