Add PartialEq to register types, fix register type conversion bugs
In order to check that the initial values for the ax5043 registers were
correct, PartialEq was added to all register types. This allowed for
reading all the registers and comparing them to the values set in
Registers::new().

Doing this discovered a handful of incorrect values, and a handful of
conversions that didn't get the correct values - mostly in the RX
Parameter Set registers.