Add aggregate registers for signal strength, rx tracking
This adds a pair of wide registers that will read their entire
respective blocks in one transaction instead of one address at a time.
This is to save on the SPI overhead of transmitting the address each
time.

It also adds the TRKAFSKDEMOD register (note the AFSK, not FSK) which is
documented in the data sheet but not the programming manual.