Add multi-chunk RX, IRQ handling
The basics of multi-chunk RX are pretty straight forward, there's some
flags that the programming manual recommends setting, and then reading
the FIFO fast enough to capture all the chunks without overflow.

The fast part though is where the details are, this patch uses the
FIFONOTEMPTY interrupt to handle this, but that requires hooking gpiod
into MIO for events, which required a patch to gpiod.

This also introduces a UI widget for displaying stuff read from the
fifo, as well as consolidating the PacketController widget.