Merge uhf/uhf_rx, add CRC checking to uhf/lband
This adds the capability for uhf to listen in RX mode but switch into TX
mode to transmit when it has data to send. This was done largely by
merging the contents of uhf_rx into uhf.

It also adds CRC checking to RX packets on both uhf and lband. In large
packet mode the ax5043 apparently just passes the crc up to the driver
to check it instead of doing checks itself. The specific CRC is supposed
to be CCITT (possibly CCITT-FALSE?) but because of AX.25 byte order
things the GENIBUS one works directly, instead of doing any conversions.