impl Index for Reg
Since Reg is a newtype wrapper for fixed arrays to satisfy the orphan
rule, it would be nice if we could just index the array directly without
going through the tuple access first.

It turns out we can! Implementing the Index trait will do exactly that,
it lets us write `reg.0[0]` as `reg[0]`. That saves a decent chunk of
line noise.