Split examples only deps into [dev-dependencies]
This is the intended use of the section, it means that any outside users
of the crate won't need the example specific stuff, like mio or clap.
Currently the crates crc and gpiod are only used in examples but I'd
eventually like to move that functionality up into the library so they
can stay where they are for now. Anyhow should only be used in the
examples but config_rpi uses it currently and needs to be refactored
onto thiserror.

[dev-dependencies] would also be used for testing libraries, once I get
to that.