Add property testing via proptest
This adds a sketch of how property, or hypothesis, testing would work
using the proptest framework. Under property testing the user describes
properties that they desire to be true and the framework will attempt to
generate test cases that prove it wrong. It generates the test cases
randomly (so, similar to fuzz testing) and when it finds a failing case
it will attempt to then find the minimal case where it fails.

This works well with the register stuff because given a register type T,
T -> into() -> tryfrom() == T is a very clear property to test.