Add proptest-derive
Proptest has many tools for making strategies out types, but in the base
crate enums are a bit clunky. Having to re-list all the variants in the
prop_oneof! macro is boilerplatey. Proptest-derive helps by introducing
derive(Arbitrary) which eliminates most of the boilerplate, at least
with simple enums.

It could be more widely used but I've not found/written an adaptor
between bitflags and proptest yet so bitflags need to be constructed by
hand still.

proptest-derive is marked as experimental still, but it seems to work
OK for my limited usecase?