Add bitorder to ChannelParameters
This consolidates the ad hoc PKTADDRCFG settings scattered throughout.
For Oresat's purposes the only LSB first channels are the beacon RX/TX.
The beacon is normal AX.25 which is common in amateur radio. On the
other hand the bulk of Oresat's data will be transmitted via CCSDS,
which is MSB first.