Fix MSK modulation setting
The configuration layer was setting the MODULATION register to FSK when
we wanted MSK. Since MSK is a specific kind of FSK, and the FSKDEV
register was being set to the correct deviation for MSK, I think it
shouldn't have a noticeable effect. I do wonder if setting it to MSK
means it automatically sets FSKDEV for us.