Add debain packaging for C3 uhf/lband
This leverages cargo-deb to do all the tedium. I had to move the uhf and
lband examples to src/bin to promote them to official binaries, along
with moving their dependencies up out of [dev-dependencies], but that
was the only change of note. I've also added the systemd units so
they're directly tracked.