Fix overflow in UHF TX
This was adding FIFOTHRESH bytes of data to a FIFOChunkTX::DATA, but
that doesn't account for the 3 byte chunk header. Removing the "chunk"
print statement in 9b8994e797 demonstrated that it was load bearing - it
slowed the TX loop down enough that it wouldn't overflow the extra 3
bytes. Removing it meant that we pretty reliably overflowed now.

I've also refactored the chunking logic. It smelled of C and slices have
literally the chunk() method. It's now obvious that PKTSTART is added to
the first chunk and PKTEND is added to the last chunk.