Update dependencies
This started as a quest to update Ratatui, who had some backwards
incompatible changes, resulting in the bulk of the code changes. While
I'm here I might as well update the rest of them, and also I tried out
using a modern tool chain.

Updating the RPI OS let me use the new toolchain, but the C3 is still on
the old stuff so no luck upgrading this time.