Unify float-like handling
There's a handful of registers that are encoded in a floating point-like
format: mantissa * 2 ^ exponent. It's be nice to have a single type that
handles that. The problem is that there's two formats: mantissa in the
high bits of width 4, and mantissa in the low bits with width 5.

This solution should handle both and converson to Reg and normal u32s.