Make FREQDEV option
None turns the frequency offset estimator off. The datasheet gives a
value to calculate this value from the bitrate and a scaling constant,
which for a given encoding just reduces to a constant, but then also it
changes based on various channel parameters? Radiolab didn't give values
consistent with this equation, so I put deriving the value on hold for
now.