Send ui data over socket
The way I've implemented the TUI is pretty processor intensive, not
great for low powered systems like the raspberry pi and especially bad
on the c3 octavo. This sends the data to generate the ui over a socket
and lets my local machine render it. I should probably also make the
rendering more efficient at some point.

Separating the ui has an add-on convenience in that the ui can be added
to the other demo apps easily without a lot of copying now. They just
need to add the bit to send stuff over the network. This can be seen in
the revival of the example rx.rs. The duplicate code was split out of
what was tui.rs and now rx.rs is purely the radio operator, and tui.rs
doesn't touch it.