Fix RxParamSets read conversion, PLLLOOP override
For RxParamSets, the parenthesis were wrong converting from the raw
value to an enum, resulting in all positions being read as the lowest
value.

For PLLLOOP, I think as a holdover of trying many things, it was being
overridden in configure_radio_rx. It should just be set when we set up
the Synthesizer.