Fix max rf offset correction, notes
While the maximum max rf offset was listed as bw/4, that seems to be
much too wide in practice. Furthermore radiolab sets it to fcarrier /
500k so that seems like a decent spot to set it.

This also adds some notes with the result of the whiteboard discussion,
and sets pktlenoffset to 0. That shouldn't have an effect but just in
case.