Move preamble patterns down one stage, use only preamble1
Assuming we wanted to use no preamble pattern and just move between
states premable1 and packet, it makes no sense then to give preamble1 a
pattern. I had originally conceptualized these as the gate for leaving a
stage, but with the way it can be configured it makes more sense to have
them as the gate for entering a stage.

Since we are using preamble 0x7E in flight, which is the SFD, then it
makes no sense to use stages 2 and 3 looking also for pattern 7E,
because they'll always be bypassed. This removes 2 and 3.