Add FOC type, fix RSSIABSTHR type
FOC or FreqOffsetCorrection is the high bit in MAXRFOFFSET. When 1 this
indicates the correction takes place at the first LO and 0 at the
second. With a bool I kept getting confused of the meaning so giving the
values names helps with that.

With RSSIABSTHR this register gets compared to RSSI, so it makes sense
that they'd be the same type and RSSI is signed, so this is probably
signed?