First pass at TOML based config
This splits the configuration out into a toml file, making it human
editable, and also if we want to change the config in flight it's much
smaller to upload.

It turns out it was much easier to include the toml parsing than I was
anticipating, I should have done this way sooner.