Convert gpiod to gpiocdev
While I'd like to upgrade to the libgpiod Rust bindings, that depends on
the 2.0 version of the library being deployed, which it's not looking
like it'll be soon in Debian. As an intermediate step gpiocdev looks at
least as feature complete as gpiod but it also implements AsRawFd so we
don't need a custom patch for it.

While doing the replacement I noticed that lband was setting pa_enable
when it shouldn't, so that's been removed.