Port carrier to rpi, add config overwrite, test synth
This was done in support of running experiments to determine exactly how
RFDIV/REFDIV/FREQA are related. It turns out that the documentation for
FREQA isn't completely correct, it should be fcarrier / (fxtal /
2^refdiv) * 2^24. Also the notblackmagic image is not correct either,
XTALDIV (perf F34) doesn't affect either SYSCLK or the PLL input.

Porting carrier was pretty straightforward, involving conditionally
compiling GPIO into it and writing a new config. To allow easy testing,
the first steps of an overwrite section were started. It's kinda clunky
currently, having to duplicate the registers. I wonder if there's an
easier way to do it.